Integrated adding chat and deleting chat into a single view.

+++ unichat_app/views.py
@@ -73,26 +73,23 @@ def user_obj_json_view(request):
 	user_obj_json = UserDetailSerializer(User.objects.get(username = username)).data
 	return JsonResponse(user_obj_json)
 
-def add_chat_view(request, **kwargs):
+def add_delete_chat_view(request, **kwargs):
 	myself = request.user
 	username = kwargs.get('username')
 	chat_user = Contact.objects.get(user = myself, contact_user = User.objects.get(username = username))
-	if ChatList.objects.filter(user = myself, chat_user = chat_user).exists():
-		chat = ChatList.objects.get(user = myself, chat_user = chat_user)
-		chat.last_active_time = current_time()
-		chat.save()
-	else:
-		ChatList.objects.create(
-			user = myself,
-			chat_user = chat_user
-		)
-	return HttpResponse({}, status = status.HTTP_201_CREATED)
-
-def delete_chat_view(request, **kwargs):
-	myself = request.user
-	username = kwargs.get('username')
-	chat_user = Contact.objects.get(user = myself, contact_user = User.objects.get(username = username))
-	if ChatList.objects.filter(user = myself, chat_user = chat_user).exists():
-		chat = ChatList.objects.get(user = myself, chat_user = chat_user)
-		chat.delete()
-	return HttpResponse({}, status = status.HTTP_200_OK)
+	if request.method == 'POST':
+		if ChatList.objects.filter(user = myself, chat_user = chat_user).exists():
+			chat = ChatList.objects.get(user = myself, chat_user = chat_user)
+			chat.last_active_time = current_time()
+			chat.save()
+		else:
+			ChatList.objects.create(
+				user = myself,
+				chat_user = chat_user
+			)
+		return HttpResponse({}, status = status.HTTP_201_CREATED)
+	elif request.method == 'DELETE':
+		if ChatList.objects.filter(user = myself, chat_user = chat_user).exists():
+			chat = ChatList.objects.get(user = myself, chat_user = chat_user)
+			chat.delete()
+		return HttpResponse({}, status = status.HTTP_200_OK)
